+++ src/main/java/com/db/crud/voting/controller/LogController.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.db.crud.voting.dto.response.LogResponse;
-import com.db.crud.voting.service.logs.LogService;
+import com.db.crud.voting.service.LogService;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
