+++ src/main/java/com/db/crud/voting/mapper/AgendaMapper.java
@@ -1,4 +1,4 @@
-package com.db.crud.voting.dto.mapper;
+package com.db.crud.voting.mapper;
 
 import java.time.LocalDateTime;
