+++ src/main/java/com/db/crud/voting/service/AgendaService.java
@@ -1,4 +1,4 @@
-package com.db.crud.voting.service.agenda;
+package com.db.crud.voting.service;
 
 import java.util.List;
