+++ src/main/java/com/db/crud/voting/service/impl/LogServiceImpl.java
@@ -1,4 +1,4 @@
-package com.db.crud.voting.service.logs;
+package com.db.crud.voting.service.impl;
 
 import java.util.List;
 
@@ -7,11 +7,12 @@ import java.time.temporal.ChronoUnit;
 
 import org.springframework.stereotype.Service;
 
-import com.db.crud.voting.dto.mapper.LogMapper;
 import com.db.crud.voting.dto.request.LogObj;
 import com.db.crud.voting.dto.response.LogResponse;
+import com.db.crud.voting.mapper.LogMapper;
 import com.db.crud.voting.model.Log;
 import com.db.crud.voting.repository.LogRepository;
+import com.db.crud.voting.service.LogService;
 
 import lombok.AllArgsConstructor;
