+++ src/test/java/com/db/crud/voting/fixture/AgendaFixture.java
@@ -2,11 +2,11 @@ package com.db.crud.voting.fixture;
 
 import java.time.LocalDateTime;
 
-import com.db.crud.voting.dto.mapper.AgendaMapper;
 import com.db.crud.voting.dto.mapper.AgendaMapperImpl;
 import com.db.crud.voting.dto.request.AgendaRequest;
 import com.db.crud.voting.dto.response.AgendaResponse;
 import com.db.crud.voting.enums.Category;
+import com.db.crud.voting.mapper.AgendaMapper;
 import com.db.crud.voting.model.Agenda;
 
 public class AgendaFixture {
