+++ src/test/java/com/db/crud/voting/fixture/UserFixture.java
@@ -1,10 +1,10 @@
 package com.db.crud.voting.fixture;
 
-import com.db.crud.voting.dto.mapper.UserMapper;
 import com.db.crud.voting.dto.mapper.UserMapperImpl;
 import com.db.crud.voting.dto.request.UserRequest;
 import com.db.crud.voting.dto.response.UserResponse;
 import com.db.crud.voting.enums.UserType;
+import com.db.crud.voting.mapper.UserMapper;
 import com.db.crud.voting.model.User;
 
 public class UserFixture {
