+++ src/test/java/com/db/crud/voting/unitary/AgendaServiceUnitaryTests.java
@@ -18,9 +18,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.db.crud.voting.dto.mapper.AgendaMapper;
-import com.db.crud.voting.dto.mapper.LogMapper;
-import com.db.crud.voting.dto.mapper.VoteMapper;
 import com.db.crud.voting.dto.request.AddVoteRequest;
 import com.db.crud.voting.dto.request.AgendaRequest;
 import com.db.crud.voting.dto.request.LogObj;
@@ -36,13 +33,16 @@ import com.db.crud.voting.fixture.AgendaFixture;
 import com.db.crud.voting.fixture.LogFixture;
 import com.db.crud.voting.fixture.UserFixture;
 import com.db.crud.voting.fixture.VoteFixture;
+import com.db.crud.voting.mapper.AgendaMapper;
+import com.db.crud.voting.mapper.LogMapper;
+import com.db.crud.voting.mapper.VoteMapper;
 import com.db.crud.voting.model.Agenda;
 import com.db.crud.voting.model.User;
 import com.db.crud.voting.repository.AgendaRepository;
 import com.db.crud.voting.repository.UserRepository;
-import com.db.crud.voting.service.agenda.AgendaServiceImpl;
-import com.db.crud.voting.service.logs.LogService;
-import com.db.crud.voting.service.user.UserServiceImpl;
+import com.db.crud.voting.service.impl.AgendaServiceImpl;
+import com.db.crud.voting.service.LogService;
+import com.db.crud.voting.service.impl.UserServiceImpl;
 
 @ExtendWith(MockitoExtension.class)
 class AgendaServiceUnitaryTests {
